requirements.txt
Flask
aiohttp
requests
python-dotenv
gunicorn